feat: only add description subject to table
Instead of adding the whole description to the table of inputs and outputs only the first line should be added.

Similar to best practice of git commit messages the description would benefit if the description will be split to subject and body.
In the first line the basic information will be placed (subject).
Followed after a blank line the body will follow.

For the action the body will then be available in the usage section only.
This will improve the readability of the table.